+++ archive/scripts/demoScripts/demoDeBridge.ts
@@ -6,7 +6,7 @@ import {
 } from '../typechain'
 import { node_url } from '../../utils/network'
 import chalk from 'chalk'
-import { config } from '../../config/debridge.json'
+import { config } from '../../../config/debridge.json'
 
 const msg = (msg: string) => {
   console.log(chalk.green(msg))
